+++ mex/common/orcid/extract.py
@@ -1,4 +1,4 @@
-from mex.common.orcid.connector import get_data_by_id, get_data_by_name
+from mex.common.orcid.connector import OrcidConnector
 from mex.common.orcid.models.person import OrcidRecord
 from mex.common.orcid.transform import map_orcid_data_to_orcid_record
 
@@ -20,7 +20,9 @@ def get_orcid_record_by_name(
     Returns:
         Orcidrecord of the matching person by name.
     """
-    orcid_data = get_data_by_name(given_names=given_names, family_name=family_name)
+    orcid_data = OrcidConnector.get_data_by_name(
+        given_names=given_names, family_name=family_name
+    )
     return map_orcid_data_to_orcid_record(orcid_data)
 
 
@@ -28,10 +30,10 @@ def get_orcid_record_by_id(orcid_id: str) -> OrcidRecord:
     """Returns Orcidrecord by UNIQUE ORCID ID.
 
     Args:
-        orcid_id: Uniqe identifier in ORCID system.
+        orcid_id: Unique identifier in ORCID system.
 
     Returns:
         Orcidrecord of the matching id.
     """
-    orcid_data = get_data_by_id(orcid_id=orcid_id)
+    orcid_data = OrcidConnector.get_data_by_id(orcid_id=orcid_id)
     return map_orcid_data_to_orcid_record(orcid_data)