+++ tests/orcid/test_extract.py
@@ -2,7 +2,6 @@ import pytest
 from requests import HTTPError
 
 from mex.common.exceptions import EmptySearchResultError, FoundMoreThanOneError
-from mex.common.orcid.connector import get_data_by_id, get_data_by_name
 from mex.common.orcid.extract import (
     get_orcid_record_by_id,
     get_orcid_record_by_name,
@@ -16,72 +15,24 @@ from mex.common.orcid.models.person import (
     OrcidPerson,
     OrcidRecord,
 )
-from tests.orcid.conftest import DUMMYDATA_JOHN, DUMMYDATA_VYVY
-
-
-def test_get_data_by_name():
-    given_names = "Vyvy "
-    family_name = "Tran Ngoc"
-    expected_result = DUMMYDATA_VYVY
-    result = get_data_by_name(given_names=given_names, family_name=family_name)
-    assert expected_result == result
-
-
-@pytest.mark.parametrize(
-    ("family_name", "given_names", "error"),
-    [
-        (
-            "Mustermann",
-            "Max",
-            pytest.raises(FoundMoreThanOneError),
-        ),
-        (
-            "Defgh",
-            "Abc",
-            pytest.raises(EmptySearchResultError),
-        ),
-    ],
-    ids=["multiple results", "empty result"],
-)
-def test_search_errors(family_name, given_names, error):
-    with error:
-        get_data_by_name(given_names=given_names, family_name=family_name)
-
-
-@pytest.mark.parametrize(
-    ("string_id", "expected", "error"),
-    [
-        ("0009-0004-3041-5706", DUMMYDATA_JOHN, None),
-        ("0009-0004-3041-576", None, pytest.raises(HTTPError)),
-        ("invalid-orcid-id", None, pytest.raises(HTTPError)),
-    ],
-    ids=["existing person", "non-existing person", "invalid characters"],
-)
-def test_get_data_by_orcid_id(string_id, expected, error) -> None:
-    if error:
-        with error:
-            get_data_by_id(orcid_id=string_id)
-    else:
-        result = get_data_by_id(orcid_id=string_id)
-        assert result == expected
 
 
 @pytest.mark.parametrize(
     ("given_names", "family_name", "expected_result", "error"),
     [
         (
-            "Vyvy",
-            "Tran Ngoc",
+            "John",
+            "Doe",
             OrcidRecord(
                 orcid_identifier=OrcidIdentifier(
-                    path="0009-0004-3041-5706",
-                    uri="https://orcid.org/0009-0004-3041-5706",
+                    path="0000-0002-1825-0097",
+                    uri="https://orcid.org/0000-0002-1825-0097",
                 ),
                 person=OrcidPerson(
                     emails=OrcidEmails(email=[]),
                     name=OrcidName(
-                        family_name=OrcidFamilyName(value="Tran Ngoc"),
-                        given_names=OrcidGivenNames(value="VyVy"),
+                        family_name=OrcidFamilyName(value="Doe"),
+                        given_names=OrcidGivenNames(value="John"),
                         visibility="public",
                     ),
                 ),
@@ -89,15 +40,15 @@ def test_get_data_by_orcid_id(string_id, expected, error) -> None:
             None,
         ),
         (
-            "Abc",
-            "Defgh",
+            "NotExistJohn",
+            "Doe",
             None,
             pytest.raises(
                 EmptySearchResultError,
             ),
         ),
         (
-            "John",
+            "Multiple",
             "Doe",
             None,
             pytest.raises(
@@ -107,6 +58,7 @@ def test_get_data_by_orcid_id(string_id, expected, error) -> None:
     ],
     ids=[],
 )
+@pytest.mark.usefixtures("mocked_orcid")
 def test_get_orcid_record_by_name(given_names, family_name, expected_result, error):
     if error:
         with error:
@@ -116,6 +68,7 @@ def test_get_orcid_record_by_name(given_names, family_name, expected_result, err
         assert result == expected_result
 
 
+@pytest.mark.usefixtures("mocked_orcid")
 @pytest.mark.parametrize(
     ("orcid_id", "expected_result", "error"),
     [
@@ -129,19 +82,15 @@ def test_get_orcid_record_by_name(given_names, family_name, expected_result, err
                 person=OrcidPerson(
                     emails=OrcidEmails(email=[]),
                     name=OrcidName(
-                        family_name=OrcidFamilyName(value="Carberry"),
-                        given_names=OrcidGivenNames(value="Josiah"),
+                        family_name=OrcidFamilyName(value="Doe"),
+                        given_names=OrcidGivenNames(value="John"),
                         visibility="public",
                     ),
                 ),
             ),
             None,
         ),
-        (
-            "0000-0001-2345-6789",
-            None,
-            pytest.raises(HTTPError),
-        ),
+        ("0000-0000-0000-0000", None, pytest.raises(HTTPError, match="404 Not Found")),
     ],
     ids=["existing id", "non-existing id"],
 )